requirements.txt
python-telegram-bot
aiosqlite
pyyaml
langchain
langchain-google-genai
langchain-core
langgraph
jinja2
beautifulsoup4
requests
numpy
openinference-instrumentation-langchain
arize-phoenix-otel